pyproject.toml
[tool.poetry]
name = "osd-dump-tools"
version = "0.1.0"
description = ""
authors = ["Daniel Callander <knifacat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
